requirements.txt
flask
authlib
requests
gunicorn
python-dotenv
pytesseract
pillow
opencv-python